Create brackets.rb
Write a function that one string as argument. The function returns 1
if the expression is correctly bracketed, otherwise it returns 0.

Symbols considered as 'brackets' are brackets '(' and ')', square brackets '['
and ']'and braces '{' and '}'. Every other symbols are simply ignored.

An opening bracket must always be closed by the good closing bracket in the
correct order. A string which not contains any bracket is considered as a
correctly bracketed string.