Create count_letter.rb
Write a function called count_letter that takes a string and returns the number of occurences of its alphabetical characters. Other characters are not counted.
The order is the order of occurence in the string. The format is in the examples.

Example 00:
Input: "abbcc"
Output: "1a, 2b, 2c"

Example 01:
Input: "My Hyze 47y 7."
Output: "1m, 3y, 1h, 1z, 1e"