Create rotone.c
Write a program that takes a string, will perform a transformation and return it. For each of the letters of the parameter string switch it by the next one in alphabetical order.
'z' becomes 'a' and 'Z' becomes 'A'. Case remains unaffected.

Example 00:

Input: "abc"
Output: "bcd"

Example 01:

Input: "AkjhZ zLKIJz , 23y "
Output: "BlkiA aMLJKa , 23z "