Create reverse_linked_list.rb
Challenge solved by my group in one of our Coding Collaborative Sessions.
Write a function that takes in the head of a linked list and returns a new list in reverse order. In this example, you can use any language that you would like. You can assume that the input will be a singly linked list.